integration_project/integration_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.4